Refactor workflow modules and action modes
